refactor(coediting): collect coedit lib into lib/coediting/ with types/svc/utils/hooks
